feat(fs-controller): enhance directory creation validation and error handling
Added input validation checks to the directory name in folder creation within fs.controller.js to ensure that only valid names are accepted. This prevents issues related to file system operations with invalid directory names. Now responds with a 400 status code and an error message when the provided name is invalid.

Additionally, implemented try-catch error handling around the directory creation block to capture any unexpected errors during the file system operation. In case of an error, it now logs the error to the console for debugging and returns a 500 status code with an error message indicating an internal server error. Also ensured the same validation logic is applied to the delete function to maintain consistency and reliability in folder operations.

+++ api/src/controllers/fs.controller.js
@@ -8,15 +8,29 @@ const Cache = require('../util/cache.util');
 module.exports.folders = {
   list: async (req, res) => res.send(await filesystem.folders().train()),
   create: (req, res) => {
-    const { name } = req.params;
-    if (!fs.existsSync(`${STORAGE.MEDIA.PATH}/train/${name}`)) {
-      fs.mkdirSync(`${STORAGE.MEDIA.PATH}/train/${name}`);
+    try {
+      const { name } = req.params;
+      if (!name || typeof name !== 'string' || !/^[a-zA-Z0-9-_]+$/.test(name)) {
+        return res.status(400).json({ error: 'Invalid directory name' });
+      }
+      if (!fs.existsSync(`${STORAGE.MEDIA.PATH}/train/${name}`)) {
+        fs.mkdirSync(`${STORAGE.MEDIA.PATH}/train/${name}`);
+      }
+      Cache.del('filters');
+      res.send({ success: true });
+    } catch (err) {
+      // Logging error to console for debugging purposes
+      console.error(err);
+
+      // Sending error response
+      res.status(500).json({ success: false, message: 'Internal server error' });
     }
-    Cache.del('filters');
-    res.send({ success: true });
   },
   delete: async (req, res) => {
     const { name } = req.params;
+    if (!name || typeof name !== 'string' || !/^[a-zA-Z0-9-_]+$/.test(name)) {
+      return res.status(400).json({ error: 'Invalid directory name' });
+    }
     if (fs.existsSync(`${STORAGE.MEDIA.PATH}/train/${name}`)) {
       fs.rmSync(`${STORAGE.MEDIA.PATH}/train/${name}`, { recursive: true });
       await resync.files();
